requirements.txt
binance_futures_connector==4.0.0
hexbytes==0.3.1
pybit==5.6.2
Pypubsub==4.0.3
python-dotenv==1.0.1
synthetix>=0.1.13
web3==6.16.0